[Add, Feat/#51]: Add IngredientData file, Ingredient Model(Struct) 추가
