--- Solidner/Models/models_dummy.swift
@@ -1,8 +1,0 @@
-//
-//  models.swift
-//  Solidner
-//
-//  Created by 황지우2 on 2023/11/06.
-//
-
-import Foundation
